Move debugging functionality to separate file
