+++ games/VentureAdventure/debugging.h
@@ -1,0 +1,12 @@
+#ifndef _DEBUGGING
+#define _DEBUGGING
+
+#include "splashkit.h"
+
+using namespace std;
+
+void process_debugging(const game_data &game, vector<string> &old_debug_message, bool &still_waiting);
+
+vector<string> get_verbose_debugging_message(const game_data &game);
+
+#endif
